Move Page subclasses to separate package
