+++ gslides_api/page/__init__.py
@@ -1,0 +1,29 @@
+from .base import BasePage, PageProperties, PageType
+from .page import (
+    Page,
+    Slide,
+    Layout,
+    Notes,
+    Master,
+    NotesMaster,
+    SlideProperties,
+    LayoutProperties,
+    NotesProperties,
+    MasterProperties,
+)
+
+__all__ = [
+    "BasePage",
+    "PageProperties",
+    "PageType",
+    "Page",
+    "Slide",
+    "Layout",
+    "Notes",
+    "Master",
+    "NotesMaster",
+    "SlideProperties",
+    "LayoutProperties",
+    "NotesProperties",
+    "MasterProperties",
+]
