+++ gslides_api/presentation.py
@@ -2,9 +2,9 @@ import copy
 import logging
 from typing import List, Optional, Dict, Any
 
-from gslides_api import Size, Dimension
+from gslides_api.page.page import Page
+from gslides_api.domain import Size
 from gslides_api.execute import api_client
-from gslides_api.page_all import Page
 from gslides_api.domain import GSlidesBaseModel
 
 logger = logging.getLogger(__name__)
